+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - 修复 Slider 在 Form 中的显示问题
 - 修复 Upload 在 onSuccess、onError 钩子无法拿到服务端返回信息的问题
 - 改善 tabs 现在支持动态更新
+- Table 新增 highlightCurrentRow 属性、新增 current-change 事件
 
 #### 非兼容性更新